requirements.txt
-i https://pypi.org/simple
pillow==6.2.1
